requirements.txt
# pip list
absl-py                 1.0.0
addict                  2.4.0
albumentations          1.1.0
antlr4-python3-runtime  4.8
anyio                   3.5.0
appdirs                 1.4.4
argon2-cffi             21.3.0
argon2-cffi-bindings    21.2.0
astor                   0.8.1
asttokens               2.0.5
astunparse              1.6.3
attrs                   21.4.0
Babel                   2.9.1
backcall                0.2.0
black                   21.4b2
bleach                  4.1.0
cachetools              4.2.2
certifi                 2021.10.8
cffi                    1.15.0
chardet                 4.0.0
charset-normalizer      2.0.10
click                   8.0.3
cloudpickle             2.0.0
colorama                0.4.4
commonmark              0.9.1
ConfigArgParse          1.5.3
contextlib2             0.6.0.post1
cycler                  0.11.0
Cython                  0.29.28
dbus-python             1.2.16
debugpy                 1.5.1
decorator               5.1.1
defusedxml              0.7.1
deprecation             2.1.0
detectron2              0.6
entrypoints             0.4
executing               0.8.2
flatbuffers             1.12
fonttools               4.28.5
freetype-py             2.2.0
future                  0.18.2
fvcore                  0.1.5.post20211023
gast                    0.3.3
geomloss                0.2.4
google-auth             1.30.0
google-auth-oauthlib    0.4.4
google-pasta            0.2.0
grpcio                  1.32.0
h5py                    2.10.0
hsluv                   5.0.2
hydra-core              1.1.1
hyperopt                0.1.2
idna                    3.3
imagecorruptions        1.1.2
imageio                 2.13.5
imageio-ffmpeg          0.4.5
imgaug                  0.4.0
importlib-metadata      4.11.1
importlib-resources     5.4.0
iniconfig               1.1.1
iopath                  0.1.9
ipdb                    0.13.9
ipykernel               6.9.1
ipython                 8.0.1
ipython-genutils        0.2.0
ipywidgets              7.6.5
jedi                    0.18.1
Jinja2                  3.0.3
joblib                  1.1.0
json-tricks             3.15.5
json5                   0.9.6
jsonschema              4.4.0
jupyter-client          7.1.2
jupyter-core            4.9.2
jupyter-packaging       0.11.1
jupyter-server          1.13.5
jupyterlab              3.2.9
jupyterlab-pygments     0.1.2
jupyterlab-server       2.10.3
jupyterlab-widgets      1.0.2
Keras-Preprocessing     1.1.2
kiwisolver              1.3.2
llvmlite                0.38.0
Markdown                3.3.6
MarkupSafe              2.1.0
matplotlib              3.5.1
matplotlib-inline       0.1.3
meshio                  5.3.0
meshplex                0.17.0
mistune                 0.8.4
mmcv                    1.4.5
mypy-extensions         0.4.3
nbclassic               0.3.5
nbclient                0.5.11
nbconvert               6.4.2
nbformat                5.1.3
nest-asyncio            1.5.4
netifaces               0.10.9
networkx                2.6.3
neural-renderer-pytorch 1.1.3
ninja                   1.10.2.3
nni                     2.2
notebook                6.4.8
npx                     0.1.0
numba                   0.55.1
numpy                   1.21.5
oauthlib                3.1.0
omegaconf               2.1.1
open3d                  0.14.1
opencv-python           4.5.5.62
opencv-python-headless  4.5.5.62
opt-einsum              3.3.0
packaging               21.3
pandas                  1.4.1
pandocfilters           1.5.0
parso                   0.8.3
pathspec                0.9.0
pexpect                 4.8.0
pickleshare             0.7.5
Pillow                  9.0.0
pip                     22.0.4
platformdirs            2.5.0
pluggy                  1.0.0
plyfile                 0.7.4
portalocker             2.3.2
prometheus-client       0.13.1
prompt-toolkit          3.0.28
protobuf                3.19.1
psutil                  5.9.0
ptyprocess              0.7.0
pure-eval               0.2.2
py                      1.11.0
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pyassimp                4.1.3
pycocotools             2.0.0
pycparser               2.21
pydot                   1.4.2
pyglet                  1.5.21
Pygments                2.11.2
PyGObject               3.36.0
pymongo                 3.11.4
PyOpenGL                3.1.0
pyparsing               3.0.6
pypng                   0.0.21
pyquaternion            0.9.9
pyrender                0.1.45
pyrsistent              0.18.1
pytest                  7.0.1
pytest-runner           5.3.1
python-dateutil         2.8.2
PythonWebHDFS           0.2.3
pytorch3d               0.6.1
pytz                    2021.3
PyWavelets              1.2.0
PyYAML                  6.0
pyzmq                   22.3.0
qudida                  0.0.4
regex                   2022.1.18
requests                2.27.1
requests-oauthlib       1.3.0
responses               0.13.3
rich                    11.2.0
rsa                     4.7.2
schema                  0.7.4
scikit-image            0.19.1
scikit-learn            1.0.2
scipy                   1.7.3
Send2Trash              1.8.0
setproctitle            1.2.2
setuptools              60.9.3
Shapely                 1.8.1.post1
simplejson              3.17.2
six                     1.16.0
sniffio                 1.2.0
stack-data              0.2.0
tabulate                0.8.9
tensorboard             2.8.0
tensorboard-data-server 0.6.1
tensorboard-plugin-wit  1.8.0
tensorboardX            2.4.1
tensorflow-estimator    2.4.0
tensorflow-gpu          2.4.0
termcolor               1.1.0
terminado               0.13.1
testpath                0.5.0
threadpoolctl           3.1.0
tifffile                2021.11.2
timm                    0.5.4
toml                    0.10.2
tomli                   2.0.1
tomlkit                 0.10.0
torch                   1.10.1+cu113
torchaudio              0.10.1+cu113
torchvision             0.11.2+cu113
tornado                 6.1
tqdm                    4.62.3
traitlets               5.1.1
transforms3d            0.3.1
trimesh                 3.9.42
typing_extensions       4.0.1
urllib3                 1.26.8
vispy                   0.9.6
wcwidth                 0.2.5
webencodings            0.5.1
websocket-client        1.2.3
websockets              9.0.1
Werkzeug                2.0.3
wget                    3.2
wheel                   0.37.1
widgetsnbextension      3.5.2
wrapt                   1.12.1
yacs                    0.1.8
yapf                    0.32.0
zipp                    3.7.0